requirements.txt
asgiref==3.8.1
click==8.1.8
Django==5.1.7
django-environ==0.12.0
djangorestframework==3.15.2
h11==0.14.0
python-dateutil==2.9.0.post0
pytz==2025.1
six==1.17.0
sqlparse==0.5.3
uvicorn==0.34.0
